+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - builtins pwd, ls, echo
+- interactive shell
+- exit
 
 [unreleased]: https://github.com/taylorhmorris/turtle/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/taylorhmorris/turtle/compare/v0.0.0...v0.0.1